+++ app/src/main/java/woowacourse/shopping/theme/Theme.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.productlist.ui.theme
+package woowacourse.shopping.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
