+++ app/src/test/java/woowacourse/shopping/feature/productlist/ProductListViewModelTest.kt
@@ -8,7 +8,6 @@ import woowacourse.shopping.feature.MainDispatcherExtension
 import woowacourse.shopping.feature.fake.FakeCartRepository
 import woowacourse.shopping.feature.fake.FakeProductRepository
 import woowacourse.shopping.feature.fake.FakeRecentProductRepository
-import woowacourse.shopping.feature.productlist.viewmodel.ProductListViewModel
 import woowacourse.shopping.fixture.TestCartContentFixture
 import woowacourse.shopping.fixture.TestProductFixture
